Core updates: i64 transition, OS leak fixes, stdlib exports
- Upgraded core `int` type in the VM from i32 to i64 to prevent float precision loss in Time tracking.
- Fixed memory leaks in `os.exec`, `os.getEnv`, and `os.chdir` by freeing strings from `valueToOwnedString`.
- Marked `os.pid`, `os.args`, and `os.setEnv` as stubs yielding `error.NotImplemented`.
- Standardized `http.fetch` to return a `[status, body]` array and removed the `Response` struct.
- Exported missing `time`, `os`, `http`, and `json` modules in `std/index.lls`.

+++ src/bytecode/value.zig
@@ -31,7 +31,7 @@ pub const ModuleObject = struct {
 pub const Value = union(enum) {
     null,
     bool: bool,
-    int: i32,
+    int: i64,
     float: f64,
     /// Pointer into the i32 heap (arrays, strings, errors, structs).
     ptr: i32,
